Add interrupt support (2)
Now it interrupts when the key "I" on the keyboard is pressed.
Next I should do interrupt on every key and send that key code to the interrupt as if it were a simple function.
Also I should add a way to switch between debug keyboard mode and interrupt mode