Add support for IR debugging
Or, simply speaking, I was bored of plugging my PS/2 keyboard in and out of my PC, so I decided to move all debugging options from keyboard to IR remote, and succeded in it.
Now it specialised for the Car MP3 remote (which was shipped with my FPGA dev board), and has 10 modes.
Modes specify, which data is showed on the 7-segment display and are changed by Ch+ and Ch- buttons.
Also you can show mode name on display by pressing Ch button and go back to data by pressing it again